requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
openai>=1.0.0
playwright>=1.40.0
ddgs>=7.0.0
numpy>=1.26.0
